@@ -1,6 +1,6 @@
 {
   "name": "@lenne.tech/nest-server",
-  "version": "11.1.14",
+  "version": "11.2.0",
   "description": "Modern, fast, powerful Node.js web framework in TypeScript based on Nest with a GraphQL API and a connection to MongoDB (or other databases).",
   "keywords": [
     "node",
@@ -66,7 +66,7 @@
   },
   "dependencies": {
     "@apollo/gateway": "2.11.2",
-    "@getbrevo/brevo": "1.0.1",
+    "@getbrevo/brevo": "3.0.1",
     "@nestjs/apollo": "13.1.0",
     "@nestjs/common": "11.1.6",
     "@nestjs/core": "11.1.6",
@@ -80,7 +80,7 @@
     "@nestjs/terminus": "11.0.0",
     "apollo-server-core": "3.13.0",
     "apollo-server-express": "3.13.0",
-    "bcrypt": "5.1.1",
+    "bcrypt": "6.0.0",
     "class-transformer": "0.5.1",
     "class-validator": "0.14.2",
     "compression": "1.8.1",
@@ -96,8 +96,8 @@
     "light-my-request": "6.6.0",
     "lodash": "4.17.21",
     "mongodb": "6.20.0",
-    "mongoose": "7.8.7",
-    "multer": "1.4.5-lts.2",
+    "mongoose": "8.19.1",
+    "multer": "2.0.2",
     "node-mailjet": "6.0.9",
     "nodemailer": "7.0.9",
     "nodemon": "3.1.10",
@@ -123,11 +123,11 @@
     "@types/cookie-parser": "1.4.9",
     "@types/ejs": "3.1.5",
     "@types/express": "4.17.21",
-    "@types/jest": "29.5.14",
+    "@types/jest": "30.0.0",
     "@types/lodash": "4.17.20",
-    "@types/multer": "1.4.12",
-    "@types/node": "22.15.17",
-    "@types/nodemailer": "6.4.17",
+    "@types/multer": "2.0.0",
+    "@types/node": "24.7.0",
+    "@types/nodemailer": "7.0.2",
     "@types/passport": "1.0.17",
     "@types/supertest": "6.0.3",
     "@typescript-eslint/eslint-plugin": "8.46.0",
@@ -143,7 +143,7 @@
     "grunt-contrib-watch": "1.1.0",
     "grunt-sync": "0.8.2",
     "husky": "9.1.7",
-    "jest": "29.7.0",
+    "jest": "30.2.0",
     "npm-watch": "0.13.0",
     "pm2": "6.0.13",
     "prettier": "3.6.2",
@@ -154,7 +154,7 @@
     "ts-morph": "25.0.1",
     "ts-node": "10.9.2",
     "tsconfig-paths": "4.2.0",
-    "typescript": "5.8.3",
+    "typescript": "5.9.3",
     "yalc": "1.0.0-pre.53"
   },
   "overrides": {
